+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 *Revisions*
 
 * Privatised underscore methods and changed old uses to using lens. ([#20](https://github.com/hjwylde/werewolf/issues/20))
+* Tidied up arbitrary instances by using `newtype`'s. ([#78](https://github.com/hjwylde/werewolf/issues/78))
 
 #### v0.4.3.1